[FSDP/DDP] Unify _cast_forward_inputs
Closes https://github.com/pytorch/pytorch/issues/96380

Differential Revision: [D46342814](https://our.internmc.facebook.com/intern/diff/D46342814/)

[ghstack-poisoned]
